requirements.txt
pyzotero>=1.5.5
python-dotenv>=1.0.0
dropbox>=11.36.0
pydub>=0.25.1
feedgen>=0.9.0
feedparser>=6.0.10
openai>=1.3.0
google-generativeai>=0.3.0
pydantic>=2.0.0
kokoro>=0.1.0
soundfile>=0.12.1
numpy>=1.24.0 